requirements.txt
Flask==2.2.3
Flask-Cors==3.0.10
Flask-JWT-Extended==4.4.4
Werkzeug==2.2.3
pymongo==4.6.3
python-dotenv==1.0.0
requests==2.32.0
openai==0.27.4
reportlab==4.0.7  # PDF生成库
pypdf2==3.0.1     # PDF处理工具
pillow==10.2.0    # ReportLab需要的图像处理库
python-alipay-sdk==3.3.0
gunicorn==23.0.0
certifi==2024.2.2
charset-normalizer==3.3.2
idna==3.6
pydantic==2.6.1
PyJWT==2.10.1
urllib3==2.2.1
blinker==1.7.0
click==8.1.7
xmltodict==0.14.2
pycryptodome==3.22.0  # 用于RSA加密

# PDF和字体相关库
future==0.18.3     # ReportLab依赖
fonttools==4.49.0  # 字体处理工具
tqdm==4.66.2       # 进度条显示

# 云存储服务相关依赖（根据需要取消注释）
# aliyun-oss2==2.15.0  # 阿里云OSS
# cos-python-sdk-v5==1.9.21  # 腾讯云COS
# boto3==1.26.84  # AWS S3
# azure-storage-blob==12.14.1  # Azure Blob Storage 